+++ playback/src/player.rs
@@ -23,11 +23,11 @@ use crate::convert::Converter;
 use crate::core::session::Session;
 use crate::core::spotify_id::SpotifyId;
 use crate::core::util::SeqGenerator;
-use crate::decoder::{AudioDecoder, AudioError, AudioPacket, PassthroughDecoder, VorbisDecoder};
+use crate::decoder::{AudioDecoder, AudioPacket, DecoderError, PassthroughDecoder, VorbisDecoder};
 use crate::metadata::{AudioItem, FileFormat};
 use crate::mixer::AudioFilter;
 
-use crate::{NUM_CHANNELS, SAMPLES_PER_SECOND};
+use crate::{MS_PER_PAGE, NUM_CHANNELS, PAGES_PER_MS, SAMPLES_PER_SECOND};
 
 const PRELOAD_NEXT_TRACK_BEFORE_END_DURATION_MS: u32 = 30000;
 pub const DB_VOLTAGE_RATIO: f64 = 20.0;
@@ -67,6 +67,8 @@ struct PlayerInternal {
     limiter_peak_sample: f64,
     limiter_factor: f64,
     limiter_strength: f64,
+
+    auto_normalise_as_album: bool,
 }
 
 enum PlayerCommand {
@@ -86,6 +88,7 @@ enum PlayerCommand {
     AddEventSender(mpsc::UnboundedSender<PlayerEvent>),
     SetSinkEventCallback(Option<SinkEventCallback>),
     EmitVolumeSetEvent(u16),
+    SetAutoNormaliseAsAlbum(bool),
 }
 
 #[derive(Debug, Clone)]
@@ -238,9 +241,10 @@ impl NormalisationData {
             return 1.0;
         }
 
-        let [gain_db, gain_peak] = match config.normalisation_type {
-            NormalisationType::Album => [data.album_gain_db, data.album_peak],
-            NormalisationType::Track => [data.track_gain_db, data.track_peak],
+        let [gain_db, gain_peak] = if config.normalisation_type == NormalisationType::Album {
+            [data.album_gain_db, data.album_peak]
+        } else {
+            [data.track_gain_db, data.track_peak]
         };
 
         let normalisation_power = gain_db as f64 + config.normalisation_pregain;
@@ -264,7 +268,11 @@ impl NormalisationData {
         }
 
         debug!("Normalisation Data: {:?}", data);
-        debug!("Normalisation Factor: {:.2}%", normalisation_factor * 100.0);
+        debug!(
+            "Calculated Normalisation Factor for {:?}: {:.2}%",
+            config.normalisation_type,
+            normalisation_factor * 100.0
+        );
 
         normalisation_factor as f64
     }
@@ -327,6 +335,8 @@ impl Player {
                 limiter_peak_sample: 0.0,
                 limiter_factor: 1.0,
                 limiter_strength: 0.0,
+
+                auto_normalise_as_album: false,
             };
 
             // While PlayerInternal is written as a future, it still contains blocking code.
@@ -346,7 +356,11 @@ impl Player {
     }
 
     fn command(&self, cmd: PlayerCommand) {
-        self.commands.as_ref().unwrap().send(cmd).unwrap();
+        if let Some(commands) = self.commands.as_ref() {
+            if let Err(e) = commands.send(cmd) {
+                error!("Player Commands Error: {}", e);
+            }
+        }
     }
 
     pub fn load(&mut self, track_id: SpotifyId, start_playing: bool, position_ms: u32) -> u64 {
@@ -406,6 +420,10 @@ impl Player {
     pub fn emit_volume_set_event(&self, volume: u16) {
         self.command(PlayerCommand::EmitVolumeSetEvent(volume));
     }
+
+    pub fn set_auto_normalise_as_album(&self, setting: bool) {
+        self.command(PlayerCommand::SetAutoNormaliseAsAlbum(setting));
+    }
 }
 
 impl Drop for Player {
@@ -415,7 +433,7 @@ impl Drop for Player {
         if let Some(handle) = self.thread_handle.take() {
             match handle.join() {
                 Ok(_) => (),
-                Err(_) => error!("Player thread panicked!"),
+                Err(e) => error!("Player thread Error: {:?}", e),
             }
         }
     }
@@ -423,7 +441,7 @@ impl Drop for Player {
 
 struct PlayerLoadedTrackData {
     decoder: Decoder,
-    normalisation_factor: f64,
+    normalisation_data: NormalisationData,
     stream_loader_controller: StreamLoaderController,
     bytes_per_second: usize,
     duration_ms: u32,
@@ -456,6 +474,7 @@ enum PlayerState {
         track_id: SpotifyId,
         play_request_id: u64,
         decoder: Decoder,
+        normalisation_data: NormalisationData,
         normalisation_factor: f64,
         stream_loader_controller: StreamLoaderController,
         bytes_per_second: usize,
@@ -467,6 +486,7 @@ enum PlayerState {
         track_id: SpotifyId,
         play_request_id: u64,
         decoder: Decoder,
+        normalisation_data: NormalisationData,
         normalisation_factor: f64,
         stream_loader_controller: StreamLoaderController,
         bytes_per_second: usize,
@@ -489,7 +509,10 @@ impl PlayerState {
         match *self {
             Stopped | EndOfTrack { .. } | Paused { .. } | Loading { .. } => false,
             Playing { .. } => true,
-            Invalid => panic!("invalid state"),
+            Invalid => {
+                error!("PlayerState is_playing: invalid state");
+                exit(1);
+            }
         }
     }
 
@@ -514,7 +537,10 @@ impl PlayerState {
             | Playing {
                 ref mut decoder, ..
             } => Some(decoder),
-            Invalid => panic!("invalid state"),
+            Invalid => {
+                error!("PlayerState decoder: invalid state");
+                exit(1);
+            }
         }
     }
 
@@ -530,7 +556,10 @@ impl PlayerState {
                 ref mut stream_loader_controller,
                 ..
             } => Some(stream_loader_controller),
-            Invalid => panic!("invalid state"),
+            Invalid => {
+                error!("PlayerState stream_loader_controller: invalid state");
+                exit(1);
+            }
         }
     }
 
@@ -543,7 +572,7 @@ impl PlayerState {
                 decoder,
                 duration_ms,
                 bytes_per_second,
-                normalisation_factor,
+                normalisation_data,
                 stream_loader_controller,
                 stream_position_pcm,
                 ..
@@ -553,7 +582,7 @@ impl PlayerState {
                     play_request_id,
                     loaded_track: PlayerLoadedTrackData {
                         decoder,
-                        normalisation_factor,
+                        normalisation_data,
                         stream_loader_controller,
                         bytes_per_second,
                         duration_ms,
@@ -561,7 +590,10 @@ impl PlayerState {
                     },
                 };
             }
-            _ => panic!("Called playing_to_end_of_track in non-playing state."),
+            _ => {
+                error!("Called playing_to_end_of_track in non-playing state.");
+                exit(1);
+            }
         }
     }
 
@@ -572,6 +604,7 @@ impl PlayerState {
                 track_id,
                 play_request_id,
                 decoder,
+                normalisation_data,
                 normalisation_factor,
                 stream_loader_controller,
                 duration_ms,
@@ -583,6 +616,7 @@ impl PlayerState {
                     track_id,
                     play_request_id,
                     decoder,
+                    normalisation_data,
                     normalisation_factor,
                     stream_loader_controller,
                     duration_ms,
@@ -592,7 +626,10 @@ impl PlayerState {
                     suggested_to_preload_next_track,
                 };
             }
-            _ => panic!("invalid state"),
+            _ => {
+                error!("PlayerState paused_to_playing: invalid state");
+                exit(1);
+            }
         }
     }
 
@@ -603,6 +640,7 @@ impl PlayerState {
                 track_id,
                 play_request_id,
                 decoder,
+                normalisation_data,
                 normalisation_factor,
                 stream_loader_controller,
                 duration_ms,
@@ -615,6 +653,7 @@ impl PlayerState {
                     track_id,
                     play_request_id,
                     decoder,
+                    normalisation_data,
                     normalisation_factor,
                     stream_loader_controller,
                     duration_ms,
@@ -623,7 +662,10 @@ impl PlayerState {
                     suggested_to_preload_next_track,
                 };
             }
-            _ => panic!("invalid state"),
+            _ => {
+                error!("PlayerState playing_to_paused: invalid state");
+                exit(1);
+            }
         }
     }
 }
@@ -679,8 +721,8 @@ impl PlayerTrackLoader {
     ) -> Option<PlayerLoadedTrackData> {
         let audio = match AudioItem::get_audio_item(&self.session, spotify_id).await {
             Ok(audio) => audio,
-            Err(_) => {
-                error!("Unable to load audio item.");
+            Err(e) => {
+                error!("Unable to load audio item: {:?}", e);
                 return None;
             }
         };
@@ -748,8 +790,8 @@ impl PlayerTrackLoader {
 
             let encrypted_file = match encrypted_file.await {
                 Ok(encrypted_file) => encrypted_file,
-                Err(_) => {
-                    error!("Unable to load encrypted file.");
+                Err(e) => {
+                    error!("Unable to load encrypted file: {:?}", e);
                     return None;
                 }
             };
@@ -767,22 +809,24 @@ impl PlayerTrackLoader {
 
             let key = match self.session.audio_key().request(spotify_id, file_id).await {
                 Ok(key) => key,
-                Err(_) => {
-                    error!("Unable to load decryption key");
+                Err(e) => {
+                    error!("Unable to load decryption key: {:?}", e);
                     return None;
                 }
             };
 
             let mut decrypted_file = AudioDecrypt::new(key, encrypted_file);
 
-            let normalisation_factor = match NormalisationData::parse_from_file(&mut decrypted_file)
-            {
-                Ok(normalisation_data) => {
-                    NormalisationData::get_factor(&self.config, normalisation_data)
-                }
+            let normalisation_data = match NormalisationData::parse_from_file(&mut decrypted_file) {
+                Ok(data) => data,
                 Err(_) => {
                     warn!("Unable to extract normalisation data, using default value.");
-                    1.0
+                    NormalisationData {
+                        track_gain_db: 0.0,
+                        track_peak: 1.0,
+                        album_gain_db: 0.0,
+                        album_peak: 1.0,
+                    }
                 }
             };
 
@@ -791,12 +835,12 @@ impl PlayerTrackLoader {
             let result = if self.config.passthrough {
                 match PassthroughDecoder::new(audio_file) {
                     Ok(result) => Ok(Box::new(result) as Decoder),
-                    Err(e) => Err(AudioError::PassthroughError(e)),
+                    Err(e) => Err(DecoderError::PassthroughDecoder(e.to_string())),
                 }
             } else {
                 match VorbisDecoder::new(audio_file) {
                     Ok(result) => Ok(Box::new(result) as Decoder),
-                    Err(e) => Err(AudioError::VorbisError(e)),
+                    Err(e) => Err(DecoderError::LewtonDecoder(e.to_string())),
                 }
             };
 
@@ -808,14 +852,17 @@ impl PlayerTrackLoader {
                         e
                     );
 
-                    if self
-                        .session
-                        .cache()
-                        .expect("If the audio file is cached, a cache should exist")
-                        .remove_file(file_id)
-                        .is_err()
-                    {
-                        return None;
+                    match self.session.cache() {
+                        Some(cache) => {
+                            if cache.remove_file(file_id).is_err() {
+                                error!("Error removing file from cache");
+                                return None;
+                            }
+                        }
+                        None => {
+                            error!("If the audio file is cached, a cache should exist");
+                            return None;
+                        }
                     }
 
                     // Just try it again
@@ -827,18 +874,20 @@ impl PlayerTrackLoader {
                 }
             };
 
-            if position_ms != 0 {
-                if let Err(err) = decoder.seek(position_ms as i64) {
-                    error!("Vorbis error: {}", err);
+            let position_pcm = PlayerInternal::position_ms_to_pcm(position_ms);
+
+            if position_pcm != 0 {
+                if let Err(e) = decoder.seek(position_pcm) {
+                    error!("PlayerTrackLoader load_track: {}", e);
                 }
                 stream_loader_controller.set_stream_mode();
             }
-            let stream_position_pcm = PlayerInternal::position_ms_to_pcm(position_ms);
+            let stream_position_pcm = position_pcm;
             info!("<{}> ({} ms) loaded", audio.name, audio.duration);
 
             return Some(PlayerLoadedTrackData {
                 decoder,
-                normalisation_factor,
+                normalisation_data,
                 stream_loader_controller,
                 bytes_per_second,
                 duration_ms,
@@ -890,7 +939,8 @@ impl Future for PlayerInternal {
                             start_playback,
                         );
                         if let PlayerState::Loading { .. } = self.state {
-                            panic!("The state wasn't changed by start_playback()");
+                            error!("The state wasn't changed by start_playback()");
+                            exit(1);
                         }
                     }
                     Poll::Ready(Err(_)) => {
@@ -954,47 +1004,67 @@ impl Future for PlayerInternal {
                     ..
                 } = self.state
                 {
-                    let packet = decoder.next_packet().expect("Vorbis error");
+                    match decoder.next_packet() {
+                        Ok(packet) => {
+                            if !passthrough {
+                                if let Some(ref packet) = packet {
+                                    match packet.samples() {
+                                        Ok(samples) => {
+                                            *stream_position_pcm +=
+                                                (samples.len() / NUM_CHANNELS as usize) as u64;
+                                            let stream_position_millis =
+                                                Self::position_pcm_to_ms(*stream_position_pcm);
 
-                    if !passthrough {
-                        if let Some(ref packet) = packet {
-                            *stream_position_pcm +=
-                                (packet.samples().len() / NUM_CHANNELS as usize) as u64;
-                            let stream_position_millis =
-                                Self::position_pcm_to_ms(*stream_position_pcm);
-
-                            let notify_about_position = match *reported_nominal_start_time {
-                                None => true,
-                                Some(reported_nominal_start_time) => {
-                                    // only notify if we're behind. If we're ahead it's probably due to a buffer of the backend and we're actually in time.
-                                    let lag = (Instant::now() - reported_nominal_start_time)
-                                        .as_millis()
-                                        as i64
-                                        - stream_position_millis as i64;
-                                    lag > Duration::from_secs(1).as_millis() as i64
+                                            let notify_about_position =
+                                                match *reported_nominal_start_time {
+                                                    None => true,
+                                                    Some(reported_nominal_start_time) => {
+                                                        // only notify if we're behind. If we're ahead it's probably due to a buffer of the backend and we're actually in time.
+                                                        let lag = (Instant::now()
+                                                            - reported_nominal_start_time)
+                                                            .as_millis()
+                                                            as i64
+                                                            - stream_position_millis as i64;
+                                                        lag > Duration::from_secs(1).as_millis()
+                                                            as i64
+                                                    }
+                                                };
+                                            if notify_about_position {
+                                                *reported_nominal_start_time = Some(
+                                                    Instant::now()
+                                                        - Duration::from_millis(
+                                                            stream_position_millis as u64,
+                                                        ),
+                                                );
+                                                self.send_event(PlayerEvent::Playing {
+                                                    track_id,
+                                                    play_request_id,
+                                                    position_ms: stream_position_millis as u32,
+                                                    duration_ms,
+                                                });
+                                            }
+                                        }
+                                        Err(e) => {
+                                            error!("PlayerInternal poll: {}", e);
+                                            exit(1);
+                                        }
+                                    }
                                 }
-                            };
-                            if notify_about_position {
-                                *reported_nominal_start_time = Some(
-                                    Instant::now()
-                                        - Duration::from_millis(stream_position_millis as u64),
-                                );
-                                self.send_event(PlayerEvent::Playing {
-                                    track_id,
-                                    play_request_id,
-                                    position_ms: stream_position_millis as u32,
-                                    duration_ms,
-                                });
+                            } else {
+                                // position, even if irrelevant, must be set so that seek() is called
+                                *stream_position_pcm = duration_ms.into();
                             }
-                        }
-                    } else {
-                        // position, even if irrelevant, must be set so that seek() is called
-                        *stream_position_pcm = duration_ms.into();
-                    }
 
-                    self.handle_packet(packet, normalisation_factor);
+                            self.handle_packet(packet, normalisation_factor);
+                        }
+                        Err(e) => {
+                            error!("PlayerInternal poll: {}", e);
+                            exit(1);
+                        }
+                    }
                 } else {
-                    unreachable!();
+                    error!("PlayerInternal poll: Invalid PlayerState");
+                    exit(1);
                 };
             }
 
@@ -1043,11 +1113,11 @@ impl Future for PlayerInternal {
 
 impl PlayerInternal {
     fn position_pcm_to_ms(position_pcm: u64) -> u32 {
-        (position_pcm * 10 / 441) as u32
+        (position_pcm as f64 * MS_PER_PAGE) as u32
     }
 
     fn position_ms_to_pcm(position_ms: u32) -> u64 {
-        position_ms as u64 * 441 / 10
+        (position_ms as f64 * PAGES_PER_MS) as u64
     }
 
     fn ensure_sink_running(&mut self) {
@@ -1058,8 +1128,8 @@ impl PlayerInternal {
             }
             match self.sink.start() {
                 Ok(()) => self.sink_status = SinkStatus::Running,
-                Err(err) => {
-                    error!("{}", err);
+                Err(e) => {
+                    error!("{}", e);
                     exit(1);
                 }
             }
@@ -1081,8 +1151,8 @@ impl PlayerInternal {
                             callback(self.sink_status);
                         }
                     }
-                    Err(err) => {
-                        error!("{}", err);
+                    Err(e) => {
+                        error!("{}", e);
                         exit(1);
                     }
                 }
@@ -1129,7 +1199,10 @@ impl PlayerInternal {
                 self.state = PlayerState::Stopped;
             }
             PlayerState::Stopped => (),
-            PlayerState::Invalid => panic!("invalid state"),
+            PlayerState::Invalid => {
+                error!("PlayerInternal handle_player_stop: invalid state");
+                exit(1);
+            }
         }
     }
 
@@ -1295,8 +1368,8 @@ impl PlayerInternal {
                         }
                     }
 
-                    if let Err(err) = self.sink.write(&packet, &mut self.converter) {
-                        error!("{}", err);
+                    if let Err(e) = self.sink.write(&packet, &mut self.converter) {
+                        error!("{}", e);
                         exit(1);
                     }
                 }
@@ -1315,7 +1388,8 @@ impl PlayerInternal {
                         play_request_id,
                     })
                 } else {
-                    unreachable!();
+                    error!("PlayerInternal handle_packet: Invalid PlayerState");
+                    exit(1);
                 }
             }
         }
@@ -1339,6 +1413,17 @@ impl PlayerInternal {
     ) {
         let position_ms = Self::position_pcm_to_ms(loaded_track.stream_position_pcm);
 
+        let mut config = self.config.clone();
+        if config.normalisation_type == NormalisationType::Auto {
+            if self.auto_normalise_as_album {
+                config.normalisation_type = NormalisationType::Album;
+            } else {
+                config.normalisation_type = NormalisationType::Track;
+            }
+        };
+        let normalisation_factor =
+            NormalisationData::get_factor(&config, loaded_track.normalisation_data);
+
         if start_playback {
             self.ensure_sink_running();
 
@@ -1353,7 +1438,8 @@ impl PlayerInternal {
                 track_id,
                 play_request_id,
                 decoder: loaded_track.decoder,
-                normalisation_factor: loaded_track.normalisation_factor,
+                normalisation_data: loaded_track.normalisation_data,
+                normalisation_factor,
                 stream_loader_controller: loaded_track.stream_loader_controller,
                 duration_ms: loaded_track.duration_ms,
                 bytes_per_second: loaded_track.bytes_per_second,
@@ -1370,7 +1456,8 @@ impl PlayerInternal {
                 track_id,
                 play_request_id,
                 decoder: loaded_track.decoder,
-                normalisation_factor: loaded_track.normalisation_factor,
+                normalisation_data: loaded_track.normalisation_data,
+                normalisation_factor,
                 stream_loader_controller: loaded_track.stream_loader_controller,
                 duration_ms: loaded_track.duration_ms,
                 bytes_per_second: loaded_track.bytes_per_second,
@@ -1423,7 +1510,10 @@ impl PlayerInternal {
                 play_request_id,
                 position_ms,
             }),
-            PlayerState::Invalid { .. } => panic!("Player is in an invalid state."),
+            PlayerState::Invalid { .. } => {
+                error!("PlayerInternal handle_command_load: invalid state");
+                exit(1);
+            }
         }
 
         // Now we check at different positions whether we already have a pre-loaded version
@@ -1439,24 +1529,30 @@ impl PlayerInternal {
             if previous_track_id == track_id {
                 let mut loaded_track = match mem::replace(&mut self.state, PlayerState::Invalid) {
                     PlayerState::EndOfTrack { loaded_track, .. } => loaded_track,
-                    _ => unreachable!(),
+                    _ => {
+                        error!("PlayerInternal handle_command_load: Invalid PlayerState");
+                        exit(1);
+                    }
                 };
 
-                if Self::position_ms_to_pcm(position_ms) != loaded_track.stream_position_pcm {
+                let position_pcm = Self::position_ms_to_pcm(position_ms);
+
+                if position_pcm != loaded_track.stream_position_pcm {
                     loaded_track
                         .stream_loader_controller
                         .set_random_access_mode();
-                    let _ = loaded_track.decoder.seek(position_ms as i64); // This may be blocking.
-                                                                           // But most likely the track is fully
-                                                                           // loaded already because we played
-                                                                           // to the end of it.
+                    if let Err(e) = loaded_track.decoder.seek(position_pcm) {
+                        // This may be blocking.
+                        error!("PlayerInternal handle_command_load: {}", e);
+                    }
                     loaded_track.stream_loader_controller.set_stream_mode();
-                    loaded_track.stream_position_pcm = Self::position_ms_to_pcm(position_ms);
+                    loaded_track.stream_position_pcm = position_pcm;
                 }
                 self.preload = PlayerPreload::None;
                 self.start_playback(track_id, play_request_id, loaded_track, play);
                 if let PlayerState::Invalid = self.state {
-                    panic!("start_playback() hasn't set a valid player state.");
+                    error!("start_playback() hasn't set a valid player state.");
+                    exit(1);
                 }
                 return;
             }
@@ -1480,11 +1576,16 @@ impl PlayerInternal {
         {
             if current_track_id == track_id {
                 // we can use the current decoder. Ensure it's at the correct position.
-                if Self::position_ms_to_pcm(position_ms) != *stream_position_pcm {
+                let position_pcm = Self::position_ms_to_pcm(position_ms);
+
+                if position_pcm != *stream_position_pcm {
                     stream_loader_controller.set_random_access_mode();
-                    let _ = decoder.seek(position_ms as i64); // This may be blocking.
+                    if let Err(e) = decoder.seek(position_pcm) {
+                        // This may be blocking.
+                        error!("PlayerInternal handle_command_load: {}", e);
+                    }
                     stream_loader_controller.set_stream_mode();
-                    *stream_position_pcm = Self::position_ms_to_pcm(position_ms);
+                    *stream_position_pcm = position_pcm;
                 }
 
                 // Move the info from the current state into a PlayerLoadedTrackData so we can use
@@ -1497,7 +1598,7 @@ impl PlayerInternal {
                     stream_loader_controller,
                     bytes_per_second,
                     duration_ms,
-                    normalisation_factor,
+                    normalisation_data,
                     ..
                 }
                 | PlayerState::Paused {
@@ -1506,13 +1607,13 @@ impl PlayerInternal {
                     stream_loader_controller,
                     bytes_per_second,
                     duration_ms,
-                    normalisation_factor,
+                    normalisation_data,
                     ..
                 } = old_state
                 {
                     let loaded_track = PlayerLoadedTrackData {
                         decoder,
-                        normalisation_factor,
+                        normalisation_data,
                         stream_loader_controller,
                         bytes_per_second,
                         duration_ms,
@@ -1523,12 +1624,14 @@ impl PlayerInternal {
                     self.start_playback(track_id, play_request_id, loaded_track, play);
 
                     if let PlayerState::Invalid = self.state {
-                        panic!("start_playback() hasn't set a valid player state.");
+                        error!("start_playback() hasn't set a valid player state.");
+                        exit(1);
                     }
 
                     return;
                 } else {
-                    unreachable!();
+                    error!("PlayerInternal handle_command_load: Invalid PlayerState");
+                    exit(1);
                 }
             }
         }
@@ -1546,17 +1649,23 @@ impl PlayerInternal {
                     mut loaded_track,
                 } = preload
                 {
-                    if Self::position_ms_to_pcm(position_ms) != loaded_track.stream_position_pcm {
+                    let position_pcm = Self::position_ms_to_pcm(position_ms);
+
+                    if position_pcm != loaded_track.stream_position_pcm {
                         loaded_track
                             .stream_loader_controller
                             .set_random_access_mode();
-                        let _ = loaded_track.decoder.seek(position_ms as i64); // This may be blocking
+                        if let Err(e) = loaded_track.decoder.seek(position_pcm) {
+                            // This may be blocking
+                            error!("PlayerInternal handle_command_load: {}", e);
+                        }
                         loaded_track.stream_loader_controller.set_stream_mode();
                     }
                     self.start_playback(track_id, play_request_id, *loaded_track, play);
                     return;
                 } else {
-                    unreachable!();
+                    error!("PlayerInternal handle_command_load: Invalid PlayerState");
+                    exit(1);
                 }
             }
         }
@@ -1662,7 +1771,9 @@ impl PlayerInternal {
             stream_loader_controller.set_random_access_mode();
         }
         if let Some(decoder) = self.state.decoder() {
-            match decoder.seek(position_ms as i64) {
+            let position_pcm = Self::position_ms_to_pcm(position_ms);
+
+            match decoder.seek(position_pcm) {
                 Ok(_) => {
                     if let PlayerState::Playing {
                         ref mut stream_position_pcm,
@@ -1673,10 +1784,10 @@ impl PlayerInternal {
                         ..
                     } = self.state
                     {
-                        *stream_position_pcm = Self::position_ms_to_pcm(position_ms);
+                        *stream_position_pcm = position_pcm;
                     }
                 }
-                Err(err) => error!("Vorbis error: {:?}", err),
+                Err(e) => error!("PlayerInternal handle_command_seek: {}", e),
             }
         } else {
             warn!("Player::seek called from invalid state");
@@ -1749,6 +1860,10 @@ impl PlayerInternal {
 
             PlayerCommand::EmitVolumeSetEvent(volume) => {
                 self.send_event(PlayerEvent::VolumeSet { volume })
+            }
+
+            PlayerCommand::SetAutoNormaliseAsAlbum(setting) => {
+                self.auto_normalise_as_album = setting
             }
         }
     }
@@ -1855,6 +1970,10 @@ impl ::std::fmt::Debug for PlayerCommand {
             PlayerCommand::EmitVolumeSetEvent(volume) => {
                 f.debug_tuple("VolumeSet").field(&volume).finish()
             }
+            PlayerCommand::SetAutoNormaliseAsAlbum(setting) => f
+                .debug_tuple("SetAutoNormaliseAsAlbum")
+                .field(&setting)
+                .finish(),
         }
     }
 }
@@ -1911,7 +2030,9 @@ struct Subfile<T: Read + Seek> {
 
 impl<T: Read + Seek> Subfile<T> {
     pub fn new(mut stream: T, offset: u64) -> Subfile<T> {
-        stream.seek(SeekFrom::Start(offset)).unwrap();
+        if let Err(e) = stream.seek(SeekFrom::Start(offset)) {
+            error!("Subfile new Error: {}", e);
+        }
         Subfile { stream, offset }
     }
 }