+++ src/main.rs
@@ -359,7 +359,7 @@ fn get_setup(args: &[String]) -> Setup {
     .optopt(
         "",
         NORMALISATION_GAIN_TYPE,
-        "Specify the normalisation gain type to use {track|album}. Defaults to album.",
+        "Specify the normalisation gain type to use {track|album|auto}. Defaults to auto.",
         "TYPE",
     )
     .optopt(